MASS_AI/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.8.0
httpx>=0.25.2
python-multipart>=0.0.6
langchain-core>=0.1.0
langgraph-checkpoint>=0.0.1
langgraph==0.0.56
supabase>=2.0.0
